Themeing and Org Changes
-added themeing to new window
-reorganized files to the proper locations
