--- PPGit/GUI/MapMaker.xaml
@@ -1,40 +1,0 @@
-﻿<Window x:Name="mapMakerFRM" x:Class="PPGit.GUI.MapMaker"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:PPGit.GUI"
-        mc:Ignorable="d"
-        Title="Map Creator" Height="463.945" Width="681.414" ResizeMode="CanMinimize" WindowStartupLocation="CenterOwner" Initialized="mapMakerFRM_Initialized" KeyDown="mapMakerFRM_KeyDown" KeyUp="mapMakerFRM_KeyUp" Closing="mapMakerFRM_Closing">
-    <Grid>
-        <Canvas x:Name="mapCVS" HorizontalAlignment="Left" Height="335" Margin="10,70,0,0" VerticalAlignment="Top" Width="644" Background="#FFF7F7F7" MouseMove="mapCVS_MouseMove" MouseLeftButtonDown="mapCVS_MouseLeftButtonDown" MouseRightButtonDown="mapCVS_MouseRightButtonDown"/>
-        <Menu HorizontalAlignment="Left" Height="32" VerticalAlignment="Top" Width="678">
-            <MenuItem x:Name="fileBTN" Header="File" Click="fileBTN_Click">
-                <MenuItem x:Name="openBTN" Header="Open" Click="openBTN_Click"/>
-                <MenuItem x:Name="saveBTN" Header="Save As..." Click="saveBTN_Click"/>
-            </MenuItem>
-            <MenuItem x:Name="backgroundLST" Header="Background" Height="31" Width="90" HorizontalAlignment="Center" VerticalAlignment="Center" Click="backgroundLST_Click">
-                <MenuItem.ContextMenu>
-                    <ContextMenu>
-                        <MenuItem x:Name="snowBTN" FontSize="12" Header="Snow" Click="snowBTN_Click"/>
-                        <MenuItem x:Name="desertBTN" FontSize="12" Header="Desert" Click="desertBTN_Click"/>
-                        <MenuItem x:Name="grassBTN" FontSize="12" Header="Grass" Click="grassBTN_Click"/>
-                        <MenuItem x:Name="waterBTN" FontSize="12" Header="Water" Click="waterBTN_Click"/>
-                    </ContextMenu>
-                </MenuItem.ContextMenu>
-            </MenuItem>
-        </Menu>
-        <Menu x:Name="iconsLST" HorizontalAlignment="Left" Height="33" Margin="0,32,0,0" VerticalAlignment="Top" Width="675">
-            <MenuItem x:Name="mountainBTN" Header="Mountains" Click="mountainBTN_Click">
-                <MenuItem.Background>
-                    <ImageBrush/>
-                </MenuItem.Background>
-            </MenuItem>
-            <MenuItem x:Name="lakeBTN" Header="Lake" Click="lakeBTN_Click"/>
-            <MenuItem x:Name="cityBTN" Header="City" Click="cityBTN_Click"/>
-            <MenuItem Header="Water Block" Click="MenuItem_Click"/>
-            <MenuItem x:Name="forestBTN" Header="Forest" Click="forestBTN_Click"/>
-        </Menu>
-
-    </Grid>
-</Window>
